Server-render hero trust stats as an integrated stat band
Replace the pill-style counters above the headline with a stat band at
the bottom of the hero copy, in the position the reference design uses
for its logo strip: large petrol numerals with small uppercase labels
above a hairline border.

Make the numbers reliable by moving the data fetch to the server:

- New getSiteStats helper queries Supabase directly and returns zeros
  when it is unconfigured or unreachable
- The home route is now a server component with 5-minute ISR
  revalidation that passes stats into the client page as props
- Numbers render in the initial HTML: no skeletons, no layout shift,
  no client fetches (previously four per visit), visible to crawlers,
  and formatted with a fixed locale to keep hydration stable
- The impact stats section reuses the same props and keeps its static
  fallbacks; the band hides entirely below minimum thresholds
- Remove the now unused hero counter components
